% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{get_info_data}
\usage{
get_info_data(mydata)
}
\arguments{
\item{mydata}{a data.frame}
}
\value{
a list with as first element the data.frame dimension and as second element the name of the columns
}
\description{
The function will provide the following information about the data.frame: dimension and column names.
}
\examples{
get_info_data(iris)
get_info_data(my_dataset)
}
