% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{get_mean_data}
\usage{
get_mean_data(mydata)
}
\arguments{
\item{mydata}{a data.frame}
}
\value{
a table with the average of all numeric variables in the data.frame
}
\description{
get_mean_data will get a table with the average of all numeric variables in your data.frame as parameter.
}
\examples{
get_mean_data(iris)
get_mean_data(my_dataset)
}
