% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_my_dataset.R
\docType{data}
\name{my_dataset}
\alias{my_dataset}
\title{my_dataset}
\format{
A data frame with 35 rows and 14 variables:
\describe{
\item{ name }{  character }
\item{ height }{  integer }
\item{ mass }{  numeric }
\item{ hair_color }{  character }
\item{ skin_color }{  character }
\item{ eye_color }{  character }
\item{ birth_year }{  numeric }
\item{ sex }{  character }
\item{ gender }{  character }
\item{ homeworld }{  character }
\item{ species }{  character }
\item{ films }{  list }
\item{ vehicles }{  list }
\item{ starships }{  list }
}
}
\source{
{dplyr}
}
\usage{
my_dataset
}
\description{
The starwars {dplyr} dataset restricted to humans (column species).
}
\keyword{datasets}
